+++ frontend/src/apis/sales/SalesAPI.tsx
@@ -4,7 +4,7 @@ import {
   METHOD_GET,
   METHOD_PATCH,
   METHOD_POST,
-} from "./Config";
+} from "../Config";
 
 const SALES_EXTENSION = "sales/sales_reconciliation";
