+++ frontend/src/apis/vendors/VendorsAPI.tsx
@@ -4,7 +4,7 @@ import {
   METHOD_GET,
   METHOD_PATCH,
   METHOD_POST,
-} from "./Config";
+} from "../Config";
 
 const VENDORS_EXTENSION = "vendors";
 // getVendors
